### 2019-06-06 #1491 @whedon
[PRE REVIEW]: OceanSpy: A Python package to facilitate ocean model data analysis and visualization
**Submitting author:** @malmans2 (<a href="http://orcid.org/0000-0001-6849-3647">Mattia Almansi</a>)
**Repository:** <a href="https://github.com/malmans2/oceanspy" target ="_blank">https://github.com/malmans2/oceanspy</a>
**Version:** v0.0.12
