### 2019-06-06 #1491 @labarba
👋  @kthyng — The submitting author suggested you as editor, and it looks right up your alley, so I assigned you. Let me know if you foresee any issues handling this. Thanks!
