### 2019-06-07 #1491 @kthyng
Yes I'll be happy to edit this!

@malmans2 do you have any suggested reviewers from [our list](https://bit.ly/joss-reviewers)?
