### 2019-06-07 #1491 @malmans2
What about edoddridge? He is familiar with Pangeo and MITgcm (OceanSpy can be tested on our MITgcm output publicly available).
