### 2019-06-07 #1491 @kthyng
@edoddridge Are you willing to do another review for JOSS? I know you did one for me pretty recently so I extra understand if you can't.

We like to have two reviewers, so I'm asking several to see if we can get two total.

@tomchor or @platipodium I see you are both new to JOSS reviewing and have relevant expertise of oceanography, modeling (though this tool can be for data too), and presumably understanding the need for good visualization. Any time and interest in reviewing this submission? We ask for reviews within 3 weeks if possible and we have [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) you can check out.
