### 2019-06-08 #1491 @tomchor
Hello, Kristen,
Yes, I'm interested in reviewing this paper and I appreciate the
invitation. My only concern is time, since I'm going to the Atmosphere and
Ocean Fluid Dyn conference at the end of the month. But if reviewing JOSS
papers take less time than traditional papers (as I imagine it does) then I
can probably finish before 3 weeks.

Thank you very much.

Best
Tomas.

On Fri, Jun 7, 2019 at 11:38 AM Kristen Thyng <notifications@github.com>
wrote:

> @edoddridge <https://github.com/edoddridge> Are you willing to do another
> review for JOSS? I know you did one for me pretty recently so I extra
> understand if you can't.
>
> We like to have two reviewers, so I'm asking several to see if we can get
> two total.
>
> @tomchor <https://github.com/tomchor> or @platipodium
> <https://github.com/platipodium> I see you are both new to JOSS reviewing
> and have relevant expertise of oceanography, modeling (though this tool can
> be for data too), and presumably understanding the need for good
> visualization. Any time and interest in reviewing this submission? We ask
> for reviews within 3 weeks if possible and we have reviewer guidelines
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html> you can
> check out.
>
