### 2019-06-13 #1491 @kthyng
Great! Let's go with @tomchor and @platipodium. @tomchor Yes I think you'll be able to finish before your conference.  If needed, you can finish after.

@edoddridge Thanks for your response!
