### 2019-06-14 malmans2/oceanspy#107 @platipodium
Add dois of software used to paper.md
Please complete your reference list in `paper.md`/`paper.bib` with dois for all software referenced (dask/xarray/xgcm/...) to continue your review evaluation https://github.com/openjournals/joss-reviews/issues/1506
