### 2019-06-14 #1506 @platipodium
Opened https://github.com/malmans2/oceanspy/issues/107 to complete doi referencing in @malmans2  `paper/paper.bib`
