### 2019-06-14 malmans2/oceanspy#107 @malmans2
@platipodium all entries other than dask have a doi now. I couldn't find a doi for dask (I also checked Zenodo), and we are using the [official citation](http://docs.dask.org/en/latest/cite.html).
