### 2019-06-14 malmans2/oceanspy#107 @platipodium
Thanks for fixing this immediately.  I don't like the dask official, as it is not a persistent identifier, but I guess we'll have to live with that.
