### 2019-06-17 #1506 @tomchor
@kthyng I'm not 100% sure if this is the appropriate behavior, but I'm pasting my review below. Most of the comments are general so I figured it wouldn't be appropriate to create an issue in the target repo for them. Please let me know if there's anything else I should do.

Best
Tomas.

Overall impressions
-------------------

This is a nice package and is likely to be helpful to scientists who do ocean modelling. The package, from what I understood, is a collection of wrappers specific to oceanography around already-existing packages. I recommend publication with minor revisions based on the comments below.

Comments on documentation checklist
-----------------------------------

Statement of need: the target audience is very well stated, but in my opinion the problems the package aims to solve could be better stated. The paper itself has those problems/needs very well explained however.

Installation instructions: although the installation instructions are very clear, there is no explicitly stated list of dependencies. Instead the dependencies are implicitly listed in a conda install command.

Automated tests: there were no automated tests that I could find in the package.

Comments on SciServer
---------------------

The SciServer part in the tutorial was a bit confusing to me. This only a suggestion, but I think It would also be good to state the differences of working from within SciServer. (I know that there is a paragraph in the beginning on the docs trying to clarify the SciServer/standalone options, but I'm not sure if everyone is familiar with SciServer.) The [flowchart](https://oceanspy.readthedocs.io/en/latest/flowchart.html), for example, could be one of the first things in the documentation (or at least in the getting started portion). Instead it is given a separate section with no context whatsoever.

Furthermore, the steps to start a SciServer didn't quite work for me. After clicking on compute, there was no option to `Run Existing Notebook`. I suspect they are outdated. It was very straightforward to figure out by myself the correct steps, but it would still be nice to have the correct instructions.

Specific comments
-----------------

- In some places (generally towards the ends of the tutorial sections) the comments tend to become scarce. An example are cells [7] through [9] in the Kogur case. A little explanation in these cases would benefit users.
- Before cell [13] in [https://oceanspy.readthedocs.io/en/latest/Tutorial.html](the tutorial) the comment should say "while this syntax adds the new variable to `od_drop`"
- I think it is really nice that in the paper (and also in the autogen docs) the calculations done by the `compute` module are mathematically defined.  However, as a scientist, it would be nice to know which numerical scheme is used to perform those computations.  I assume it is the default functionality of `xarray`, which uses specific `numpy` functions. A simple reference to the `xarray` function would suffice (to avoid having to go through code to figure it out).
