### 2019-06-17 malmans2/oceanspy#109 @ThomasHaine
Sounds good to me. The other issues are easy to fix. Let me know if you need help!
