### 2019-06-17 malmans2/oceanspy#109 @asiddi24
Are we supposed to wait for both reviews to come before making changes ?
