### 2019-06-17 malmans2/oceanspy#109 @malmans2
I don't think so. Let's open a new pull request. I'll ask the editor if we can just merge it.
