### 2019-06-17 malmans2/oceanspy#109 @renskegelderloos
Sounds good! These are some good suggestions. Let me know if I can help/check edits etc.

Renske
