### 2019-06-17 be81d0e
removed flowchart and added to quick start
