+++ docs/flowchart.rst
@@ -1,8 +1,0 @@
-.. _flowchart:
-
-Flowchart
----------
-
-.. image:: ./_static/flowchart.jpg
-   :alt: Flowchart
-
