+++ docs/index.rst
@@ -8,7 +8,6 @@
    installation
    sciserver
    datasets
-   flowchart
 
 .. toctree::
    :maxdepth: 1
