### 2019-06-17 malmans2/oceanspy#109 @asiddi24
I have made a pull request addressing points 1, 2, 4 and 6. See if we that works.
