### 2019-06-17 malmans2/oceanspy#109 @malmans2
Thanks! I'll be on it probably tomorrow. I think I can take care of the other points.
