### 2019-06-18 #1506 @platipodium
I am happy with merging the PR, us rebuilding docs is part of the job description I guess.
