+++ docs/installation.rst
@@ -4,37 +4,48 @@
 Installation
 ============
 
-If you would like to use OceanSpy for your own datasets and run it on a local machine or server, you will need to download and install it first. Use the following 3 steps to install OceanSpy.
+If you would like to use OceanSpy for your own datasets and run it on a local machine or server, you will need to download and install it first.
+
+Required dependencies
+---------------------
+
+* Python 3.6, 3.7
+* dask_
+* xarray_
+* xgcm_
+
+Optional dependencies
+---------------------
+
+For optimal performance and to enable all OceanSpy features, it is highly recommended that you install the following dependencies:
+
+* bottleneck_  
+* Cartopy_  
+* Dask.distributed_  
+* ESMPy_  
+* FFmpeg_  
+* GeoPy_  
+* intake_xarray_  
+* IPython_  
+* netCDF4_  
+* tqdm_  
+* xESMF_  
+* xmitgcm_  
+
+Step-by-step instructions
+-------------------------
 
 1. First, download and install Miniconda_ or Anaconda_.
 
-2. Second, OceanSpy and its dependencies need to be installed. The easiest way to install OceanSpy's dependencies is to use the conda-forge channel. 
-
-  For optimal performance, OceanSpy requires the latest version of the following python dependencies to be installed : 
-
-  - Dask.distributed_
-  - bottleneck_
-  - netCDF4_
-  - xarray_
-  - Cartopy_
-  - ESMPy_
-  - FFmpeg_
-  - intake_xarray_
-  - GeoPy_
-  - xgcm_
-  - xESMF_
-  - IPython_
-  - tqdm_
-
-  To install OceanSpy and the above mentioned dependencies, open a terminal, then run the following commands:
+2. The easiest way to install OceanSpy and the above mentioned dependencies is to use the conda-forge channel. Open a terminal, then run the following commands:
 
   .. code-block:: bash
     
     $ conda update conda
     $ conda config --set channel_priority strict
     $ conda config --prepend channels conda-forge
-    $ conda install -y oceanspy dask distributed bottleneck netCDF4 "xarray>=0.11.3" cartopy esmpy ffmpeg intake-xarray geopy "xgcm>=0.2" xesmf Ipython tqdm
-    $ pip install "xmitgcm>=0.3" ffmpeg
+    $ conda install -y oceanspy dask distributed bottleneck netCDF4 "xarray>=0.11.3" cartopy esmpy ffmpeg intake-xarray geopy "xgcm>=0.2" xesmf Ipython tqdm "xmitgcm>=0.3"
+    $ pip install ffmpeg
 
   The commands above install the latest stable release of OceanSpy along with its dependencies.
 
@@ -63,19 +74,25 @@ If you would like to use OceanSpy for your own datasets and run it on a local ma
 
 		$ conda activate Oceanography
 
+Test suite
+----------
+Step-by-step instructions on how to run the test suite are available in :ref:`code`.
+
+.. _xarray: http://xarray.pydata.org
+.. _dask: https://dask.org
+.. _xgcm: https://xgcm.readthedocs.io
 .. _Anaconda: https: //www.anaconda.com/
-.. _Miniconda: https: //docs.conda.io/en/latest/miniconda.html
-.. _Dask.distributed : http://distributed.dask.org/en/latest/
-.. _bottleneck : https://github.com/kwgoodman/bottleneck
-.. _netCDF4 : https://unidata.github.io/netcdf4-python/netCDF4/index.html
-.. _xarray : http://xarray.pydata.org/en/stable/
[…]
