### 2019-06-19 d9a7441
flowchart rewording
