### 2019-06-19 #1506 @malmans2
@tomchor, thanks for your quick and helpful review! I think we've been able to address all your points. The new documentation is already online.

> Statement of need: the target audience is very well stated, but in my opinion the problems the package aims to solve could be better stated. The paper itself has those problems/needs very well explained however.

We edited the github README, which is also the front page of the documentation: https://oceanspy.readthedocs.io

> Installation instructions: although the installation instructions are very clear, there is no explicitly stated list of dependencies. Instead the dependencies are implicitly listed in a conda install command.

Required and optional dependencies are now explicitly stated [here](https://oceanspy.readthedocs.io/en/latest/installation.html).

> Automated tests: there were no automated tests that I could find in the package.

We have automated tests and they cover 98% of the code. Instructions on how to run and implement tests are available [here](https://oceanspy.readthedocs.io/en/latest/contributing.html#contributing-to-the-code). We also added a section in the [installation](https://oceanspy.readthedocs.io/en/latest/installation.html) page to make it more clear. You can check [continuous integration](https://travis-ci.org/malmans2/oceanspy) and [code coverage](https://codecov.io/github/malmans2/oceanspy?branch=master) by clicking on the stickers in the front page.

> The SciServer part in the tutorial was a bit confusing to me. This only a suggestion, but I think It would also be good to state the differences of working from within SciServer. (I know that there is a paragraph in the beginning on the docs trying to clarify the SciServer/standalone options, but I'm not sure if everyone is familiar with SciServer.) The flowchart, for example, could be one of the first things in the documentation (or at least in the getting started portion). Instead it is given a separate section with no context whatsoever.

We revised the [quick start](https://oceanspy.readthedocs.io/en/latest/quick.html) and [SciServer](https://oceanspy.readthedocs.io/en/latest/sciserver.html) sections to make it more clear, and the flowchart is now in the quick start section. We also added in bold  "How to use OceanSpy?" in the front page.

> Furthermore, the steps to start a SciServer didn't quite work for me. After clicking on compute, there was no option to Run Existing Notebook. I suspect they are outdated. It was very straightforward to figure out by myself the correct steps, but it would still be nice to have the correct instructions.

We had a typo in the [SciServer Job](https://oceanspy.readthedocs.io/en/latest/sciserver.html#compute-jobs) section that originated the confusion. Point 3 now says: click on `compute Jobs`, which is the SciServer app that allows to run jobs asynchronously. The steps in quick-start, which use the interactive domain, should be up-to-date.

> In some places (generally towards the ends of the tutorial sections) the comments tend to become scarce. An example are cells [7] through [9] in the Kogur case. A little explanation in these cases would benefit users.

All code cells are now preceded by markdown comments.

> Before cell [13] in [https://oceanspy.readthedocs.io/en/latest/Tutorial.html](the tutorial) the comment should say "while this syntax adds the new variable to od_drop"

Fixed!

> I think it is really nice that in the paper (and also in the autogen docs) the calculations done by the compute module are mathematically defined. However, as a scientist, it would be nice to know which numerical scheme is used to perform those computations. I assume it is the default functionality of xarray, which uses specific numpy functions. A simple reference to the xarray function would suffice (to avoid having to go through code to figure it out).

OceanSpy is mainly targeting models that use finite volume methods. We tested OceanSpy using MITgcm simulations, so we used the MITgcm schemes as reference.
Here is an example: To compute the vertical component of the relative vorticity we use the OceanSpy's `curl` function. The vorticity computed by the model online and by OceanSpy agree to machine precision. [OceanSpy's curl](https://oceanspy.readthedocs.io/en/latest/generated/oceanspy.compute.curl.html) has a "Numerical Method" reference that points to the MITgcm documentation. Since [OceanSpy's vertical_relative_vorticity](https://oceanspy.readthedocs.io/en/latest/generated/oceanspy.compute.vertical_relative_vorticity.html) uses the curl function, `curl` is listed under "See Also" of `vertical_relative_vorticity` (it's also a link).
