### 2019-06-20 malmans2/oceanspy#109 @malmans2
The reviewer is making a good point [here](https://github.com/openjournals/joss-reviews/issues/1506#issuecomment-503770089). I suggest we add a note in the front page. This is what I'm thinking:
 "OceanSpy has been developed and tested using MITgcm output. However, it has been thought to be compatible with any ocean general circulation model. OceanSpy's structure allows to easily implement additional model-specific features, such as numerical schemes for vector calculus, budget closures, and equations of state. We are actively seeking feedback from different ocean model communities ([feedback submission](https://github.com/malmans2/oceanspy/issues))."
Thoughts? @ThomasHaine @renskegelderloos @asiddi24 @hooteoos-waltz
