### 2019-06-20 malmans2/oceanspy#109 @ThomasHaine
Yes, good point. I suggest this text:

"OceanSpy has been developed and tested using MITgcm output. However, it is designed to work with any (structured grid) ocean general circulation model. OceanSpy's architecture allows to easily implement model-specific features, such as different grids, numerical schemes for vector calculus, budget closures, and equations of state. We actively seek input and contributions from users of other ocean models (feedback submission<https://github.com/malmans2/oceanspy/issues>)."
