### 2019-06-25 #1506 @kthyng
Thanks for everyone's effort so far! Sorry for my slow response; I've been traveling.

@tomchor Are there any lingering issues you'd like to see addressed? I see one checkbox still unchecked.

@platipodium How are things from your perspective?
