### 2019-06-25 #1506 @platipodium
I have so far failed to successfully install the software on my local system, neither  via conda forge nor on my system python installation, that's why I did not check functionality.
