### 2019-06-25 #1506 @malmans2
@platipodium could you please tell me a little more about the issues you are encountering?
In my experience the following commands usually solve all problems:
```
$ conda config --set channel_priority strict
$ conda config --prepend channels conda-forge
```

They basically just edit you `.condarc`, which should look like:

```console
channel_priority: strict
channels:
  - conda-forge
  - defaults
```

If your `.condarc` has been edited before and looks different, or if you are using a very old version of conda, this could be another source of problems.
