### 2019-07-03 #1506 @malmans2
@platipodium, just checking if you are still stuck on the installation. I'm able to install on our machines and the [continuous integration](https://travis-ci.org/malmans2/oceanspy) is also successfully creating the environment, so I would need more details to get it fixed (e.g., log, OS, ...).
