### 2019-07-03 #1506 @platipodium
Thanks for checking on me.  I have finally been able to overcome the issues I faced, all of which were related to permissions gone wrong in my `conda` installation.  Nothing to worry about in your package.   But something we too often forget with CI and package management, that also those systems are far from perfect and working for everyone.   I have not yet gone through the steps to replicating your demo functionality on my local system, but that will be done soon.  Sorry for the delay.
