### 2019-07-03 #1506 @malmans2
Good news, thanks! I was worried about problems related to the OS since we mainly test and develop on linux machines.
