### 2019-07-04 #1506 @platipodium
@malmans2  I have now tested the claimed funtcionality, also on my local system, @kthyng I still can't tick my boxes (but I would if I could).  So I guess the review is complete.
