### 2019-07-05 #1506 @kthyng
Hi @malmans2! Can you clarify which ocean models OceanSpy will work with in your paper? After that, please make a Zenodo archive, and report the DOI in this thread.
