### 2019-07-05 malmans2/oceanspy#109 @malmans2
@ThomasHaine @renskegelderloos @asiddi24 @hooteoos-waltz
Last task: the editor is asking to clarify which ocean models OceanSpy will work with in the [paper](https://github.com/openjournals/joss-papers/blob/joss.01506/joss.01506/10.21105.joss.01506.pdf).
I suggest we add the following line at the end of the Summary section:
"It is designed to work with structured grid ocean circulation models that use a finite volume approach to discretize in space."
What do you think?
