### 2019-07-05 malmans2/oceanspy#109 @renskegelderloos
Only C-grid, but otherwise ok.
