### 2019-07-05 malmans2/oceanspy#109 @malmans2
Right, let's be more specific. I'll add the Arakawa citation then. So it would be:

It is designed to work with structured C-grid ([Arakawa and Lamb, 1977](https://www.sciencedirect.com/science/article/pii/B9780124608177500094)) ocean circulation models that use a finite volume approach to discretize in space.

If it works for @ThomasHaine, I'll create the doi for the paper.
