### 2019-07-05 malmans2/oceanspy#109 @ThomasHaine
Is the finite volume part necessary?  Presumably, it would be relatively straightforward to adjust for A or B grid models in future?
