### 2019-07-05 malmans2/oceanspy#109 @malmans2
I originally thought to add the finite volume bit because it tells about the numerics currently used in compute, and it includes most of the GCM that use C-grid. Now I'm thinking that maybe we could just add a separate section in the paper (e.g., Model compatibility), where we include the same sentences we have on the front page with a reference to the the MITgcm:
### Model compatibility
OceanSpy has been developed and tested using output of the Massachusets Institute of Technology General Circulation Model (MITgcm; [Marshall et al. 1997](https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/96JC02775)). However, it is designed to work with any (structured grid) ocean general circulation model. OceanSpy’s architecture allows to easily implement model-specific features, such as different grids, numerical schemes for vector calculus, budget closures, and equations of state.
