+++ paper/paper.bib
@@ -55,3 +55,14 @@
   url       = {http://www.cambridge.org/9780521768436},
 	doi       = {10.1017/9781139015721}
 }
+
+@article{Marshall97,
+author = {Marshall, John and Adcroft, Alistair and Hill, Chris and Perelman, Lev and Heisey, Curt},
+title = {A finite-volume, incompressible Navier Stokes model for studies of the ocean on parallel computers},
+journal = {Journal of Geophysical Research: Oceans},
+volume = {102},
+number = {C3},
+pages = {5753-5766},
+doi = {10.1029/96JC02775},
+year = {1997}
+}
