+++ HISTORY.rst
@@ -4,15 +4,16 @@
 History
 =======
 
-0.0.12 (2019-05-24)
--------------------
+v0.1 (2019-07-06)
+-----------------
 
-OceanSpy is now ready to be disseminated and submitted to JOSS.
-Version 0.1 will be released after the peer review process.
-`Mattia Almansi`_, `Renske Gelderloos`_, `Tom Haine`_, `Atousa Saberi`_, and `Ali Siddiqui`_ contributed to the development of OceanSpy.
+Initial release published in the `Journal of Open Source Software`_.
+
+`Mattia Almansi`_, `Renske Gelderloos`_, `Tom Haine`_, `Atousa Saberi`_, and `Ali Siddiqui`_ contributed to the development.
 
 .. _`Mattia Almansi`: https://github.com/malmans2
 .. _`Renske Gelderloos`: https://github.com/renskegelderloos
 .. _`Tom Haine`: https://github.com/ThomasHaine
 .. _`Atousa Saberi`: https://github.com/hooteoos-waltz
 .. _`Ali Siddiqui`: https://github.com/asiddi24
+.. _`Journal of Open Source Software`: https://joss.theoj.org
