+++ setup.py
@@ -45,6 +45,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/malmans2/oceanspy',
-    version='0.0.12',
+    version='0.1.0',
     zip_safe=False,
 )
