### 2019-07-06 c67f634
Update HISTORY.rst
