### 2019-07-06 #1506 @malmans2
> Hi @malmans2! Can you clarify which ocean models OceanSpy will work with in your paper? After that, please make a Zenodo archive, and report the DOI in this thread.

Hi @kthyng, I think we're all set. We've updated the paper and added a reference. Before creating the doi I've released v0.1.0, so I think we need to [inform whedon](https://joss.readthedocs.io/en/latest/whedon.html#changing-the-software-version).
Here is the doi: 10.5281/zenodo.3270646
