### 2019-07-06 #1506 @kthyng
@malmans2 excellent! You may be getting some notes in the future as I encourage people I know to try out this package!

I'll have whedon update several things and feel free to let me know if you think I've missed something.

First though, do you want to keep the version as 0.1 instead of 1? Publishing this paper could be a good reason to bump up the version all the way to a whole number :)
