### 2019-07-06 #1506 @kthyng
(obviously completely up to you, just a suggestion)
