### 2019-07-06 #1506 @malmans2
Thanks a lot for the feedback and helping disseminate OceanSpy!
I would keep 0.1 for now. I think we have a few tasks before version 1:
- Get more feedback and contributions from other groups (both observational and modeling)
- Test/adapt OceanSpy on different models
- Fully implement the particle code
- Upload LLC4320 on SciServer and test/improve OceanSpy's performances on high-res global datasets.
